src/main/webapp/app/entities/disponibility/route/disponibility-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IDisponibility, Disponibility } from '../disponibility.model';
import { DisponibilityService } from '../service/disponibility.service';

import { DisponibilityRoutingResolveService } from './disponibility-routing-resolve.service';

describe('Disponibility routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: DisponibilityRoutingResolveService;
  let service: DisponibilityService;
  let resultDisponibility: IDisponibility | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(DisponibilityRoutingResolveService);
    service = TestBed.inject(DisponibilityService);
    resultDisponibility = undefined;
  });

  describe('resolve', () => {
    it('should return IDisponibility returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDisponibility = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDisponibility).toEqual({ id: 123 });
    });

    it('should return new IDisponibility if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDisponibility = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultDisponibility).toEqual(new Disponibility());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Disponibility })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultDisponibility = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultDisponibility).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
